requirements.txt
numpy==1.23.4
pandas==1.5.1
pdoc3==0.10.0
coverage==6.5.0
plotly==5.11.0
scipy==1.9.3
matplotlib==3.6.3
PyYAML==6.0
scikit-learn==1.3.0
xgboost==1.7.6
coloredlogs==15.0.1
seaborn==0.12.2
pyyaml==6.0